Move B0's ocaml support to its own lib.
The use of compiler libs trips brzo --top which is annoying.
It's likely a bad idea for `B0` to rely on compiler libs. compiler
libs being just used for getting info in the cmis for brzo's
best-effort strategy we should likely move that back to b0 at
which point we could reintegrate ocaml's support in src-care.