Fix misleading "never became ready" errors.
We need to keep track of files that are made ready via Memo.file_ready.
For now we simply keep this in memo values. Adding a dedicated case in
Op.t could be an option (but seems heavy).

We also didn't add these to the log for now, they would be needed to
be able to use Op.find_aggregate_error on a log without misleading.

We might consider doing some of these things in the future.